Move testRegex out of Dialog class
